+++ .cursorrules
@@ -25,8 +25,6 @@ You are an expert in React, Shadcn UI, GraphQL, Langchain, Langgraph, and Tailwi
   │   ├── /lib
   │   │   ├── /utils
   │   │   └── /supabase
-  │   ├── /routes
-  │   │   └── /(route-components)
   │   ├── /types
   │   ├── /assets
   │   ├── App.tsx
