+++ src/agents/queryBuilder.ts
@@ -6,7 +6,7 @@ import {
 import { BaseMessage, HumanMessage } from "@langchain/core/messages";
 import { GraphQLSchema, parse, validate, buildClientSchema } from "graphql";
 import { llmModel } from "../lib/llmClient.js";
-import { QUERY_BUILDER_PROMPT_TEMPLATE } from "../prompts/agent/query-builder.js";
+import { QUERY_BUILDER_PROMPT_TEMPLATE } from "../prompts/agent/queryBuilder.js";
 
 export interface QueryBuilderState {
   messages: BaseMessage[];
