+++ src/prompts/agent/queryBuilder.ts
