+++ src/prompts/agent/responseFormatter.ts
