+++ src/prompts/developer/queryGeneration.ts
